# api/openapi.yaml
openapi: "3.1.0"
info:
  title: "Kinetic v3 API"
  version: "0.1.0"
  description: |
    REST API for the Kinetic v3 Rocket League esports league management platform.

    ## Authentication
    Most read endpoints are publicly accessible. Write and sensitive operations
    require authentication via either a session cookie (set by the OAuth login flow)
    or an HTTP Bearer token (API token issued via `/v1/api-tokens`).

    ## Permission model
    Protected routes enforce a resource/action pair checked against the caller's
    roles. Available resources are defined in `authz.resources` and include
    `league`, `franchise`, `club`, `team`, `player`, `roster_membership`,
    `role_assignment`, `queue`, `queue_entry`, `queue_ban`, `scrim`,
    `scrim_promotion`, `result_submission`, `result_override`, `replay_evidence`,
    `platform_account_link`, `player_rating`, `rating_adjustment`,
    `matchmaking_decision`, `exception_ticket`, `season`, `schedule_group`,
    `fixture`, `match`, `eligibility`, `player_stat`, `skill_group`,
    `skill_group_transition`, `organization_config`, `api_token`, and
    `player_notification`.
servers:
  - url: "http://localhost:8080"
    description: "Local development server"
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: kinetic_session
      description: "Signed session token set by the OAuth login flow."
    bearerToken:
      type: http
      scheme: bearer
      description: "Long-lived API token issued via POST /v1/api-tokens."
  schemas:
...
    Season:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        slug:
          type: string
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - slug
        - isActive
        - createdAt
    ScheduleGroup:
      type: object
      properties:
        id:
          type: integer
          format: int64
        seasonId:
          type: integer
          format: int64
        name:
          type: string
        sequence:
          type: integer
          format: int32
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - seasonId
        - name
        - sequence
        - isActive
        - createdAt
    Fixture:
      type: object
      properties:
        id:
          type: integer
          format: int64
        scheduleGroupId:
          type: integer
          format: int64
        homeClubId:
          type: integer
          format: int64
        awayClubId:
          type: integer
          format: int64
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - scheduleGroupId
        - homeClubId
        - awayClubId
        - isActive
        - createdAt
    Match:
      type: object
      properties:
        id:
          type: integer
          format: int64
        fixtureId:
          type: integer
          format: int64
        homeTeamId:
          type: integer
          format: int64
        awayTeamId:
          type: integer
          format: int64
        state:
          type: string
        scheduledFor:
          type: string
          format: date-time
          nullable: true
        homeTimeRatifiedAt:
          type: string
          format: date-time
          nullable: true
        awayTimeRatifiedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - fixtureId
        - homeTeamId
        - awayTeamId
        - state
        - createdAt

    # ── Skill Groups ───────────────────────────────────────────────────────────
...
paths:
...
  /v1/seasons:
    get:
      operationId: listSeasons
      summary: "List all seasons"
      tags:
        - scheduling
      responses:
        "200":
          description: "All seasons."
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Season"
        "503":
          description: "Hierarchy store unavailable."
    post:
      operationId: createSeason
      summary: "Create a new season"
      tags:
        - scheduling
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                slug:
                  type: string
              required:
                - name
                - slug
      responses:
        "201":
          description: "Created season."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Season"
        "400":
          description: "Invalid input."
        "409":
          description: "Conflict."
        "503":
          description: "Hierarchy store unavailable."
  /v1/schedule-groups:
    get:
      operationId: listScheduleGroups
      summary: "List all schedule groups (weeks/rounds)"
      tags:
        - scheduling
      responses:
        "200":
          description: "All schedule groups."
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ScheduleGroup"
        "503":
          description: "Hierarchy store unavailable."
    post:
      operationId: createScheduleGroup
      summary: "Create a schedule group within a season"
      tags:
        - scheduling
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                seasonId:
                  type: integer
                  format: int64
                name:
                  type: string
                sequence:
                  type: integer
                  format: int32
              required:
                - seasonId
                - name
                - sequence
      responses:
        "201":
          description: "Created schedule group."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ScheduleGroup"
        "400":
          description: "Invalid input."
        "409":
          description: "Conflict."
        "503":
          description: "Hierarchy store unavailable."
  /v1/fixtures:
    get:
      operationId: listFixtures
      summary: "List all fixtures"
      tags:
        - scheduling
      responses:
        "200":
          description: "All fixtures."
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Fixture"
        "503":
          description: "Hierarchy store unavailable."
    post:
      operationId: createFixture
      summary: "Create a fixture (club vs club matchup in a schedule group)"
      tags:
        - scheduling
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                scheduleGroupId:
                  type: integer
                  format: int64
                homeClubId:
                  type: integer
                  format: int64
                awayClubId:
                  type: integer
                  format: int64
              required:
                - scheduleGroupId
                - homeClubId
                - awayClubId
      responses:
        "201":
          description: "Created fixture."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fixture"
        "400":
          description: "Invalid input."
        "409":
          description: "Conflict."
        "503":
          description: "Hierarchy store unavailable."
  /v1/fixtures/{id}:
    get:
      operationId: getFixture
      summary: "Get a single fixture by ID"
      tags:
        - scheduling
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: "The fixture."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Fixture"
        "400":
          description: "Invalid fixture ID."
        "404":
          description: "Fixture not found."
        "503":
          description: "Hierarchy store unavailable."
  /v1/matches:
    get:
      operationId: listMatches
      summary: "List all matches"
      tags:
        - scheduling
      responses:
        "200":
          description: "All matches."
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Match"
        "503":
          description: "Hierarchy store unavailable."
    post:
      operationId: createMatch
      summary: "Create a match within a fixture"
      tags:
        - scheduling
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                fixtureId:
                  type: integer
                  format: int64
                homeTeamId:
                  type: integer
                  format: int64
                awayTeamId:
                  type: integer
                  format: int64
                state:
                  type: string
                scheduledFor:
                  type: string
                  format: date-time
                  nullable: true
                homeTimeRatifiedAt:
                  type: string
                  format: date-time
                  nullable: true
                awayTimeRatifiedAt:
                  type: string
                  format: date-time
                  nullable: true
              required:
                - fixtureId
                - homeTeamId
                - awayTeamId
                - state
      responses:
        "201":
          description: "Created match."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Match"
        "400":
          description: "Invalid input."
        "409":
          description: "Conflict."
        "503":
          description: "Hierarchy store unavailable."

  # ── Skill Groups ───────────────────────────────────────────────────────────
...